[sub/estimate-1] 수의사 관련 모듈 코드 리팩토링 및 persistence 모듈 코드 리팩토링 완료
