+++ daengle-persistence/src/main/java/ddog/persistence/adapter/GroomerRepository.java
@@ -13,13 +13,6 @@ public class GroomerRepository implements GroomerPersist {
     private final GroomerJpaRepository groomerJpaRepository;
 
     @Override
-    public String getAddressById(Long accountId) {
-        return groomerJpaRepository.findByAccountId(accountId)
-                .orElseThrow(() -> new UserException(UserExceptionType.USER_NOT_FOUND))
-                .getAddress();
-    }
-
-    @Override
     public Groomer getGroomerByAccountId(Long accountId) {
         return groomerJpaRepository.getByAccountId(accountId)
                 .toModel();
