+++ daengle-persistence/src/main/java/ddog/persistence/adapter/UserRepository.java
@@ -21,7 +21,7 @@ public class UserRepository implements UserPersist {
     @Override
     public User findByAccountId(Long accountId) {
         return userJpaRepository.findByAccountId(accountId)
-                .orElseThrow(() -> new UserException(UserExceptionType.USER_NOT_FOUND))
+                .orElseThrow(() -> new RuntimeException("user not found"))
                 .toModel();
     }
