+++ daengle-persistence/src/main/java/ddog/persistence/jpa/entity/GroomingEstimateJpaEntity.java
@@ -1,9 +1,9 @@
 package ddog.persistence.jpa.entity;
 
-import ddog.account.enums.Weight;
-import ddog.daengleserver.domain.estimate.EstimateStatus;
-import ddog.daengleserver.domain.estimate.GroomingEstimate;
-import ddog.daengleserver.domain.estimate.Proposal;
+import ddog.domain.estimate.EstimateStatus;
+import ddog.domain.estimate.GroomingEstimate;
+import ddog.domain.estimate.Proposal;
+import ddog.domain.pet.Weight;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
