+++ daengle-persistence/src/main/java/ddog/persistence/jpa/repository/VetJpaRepository.java
@@ -1,6 +1,6 @@
 package ddog.persistence.jpa.repository;
 
-import ddog.daengleserver.infrastructure.po.VetJpaEntity;
+import ddog.persistence.jpa.entity.VetJpaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
