+++ daengle-persistence/src/main/java/ddog/persistence/port/AccountPersist.java
@@ -4,12 +4,8 @@ import ddog.domain.account.Account;
 import ddog.domain.account.Role;
 
 public interface AccountPersist {
-
     boolean checkExistsAccountBy(String email, Role role);
-
     Account save(Account account);
-
     Account findById(Long accountId);
-
     Account findAccountByEmailAndRole(String email, Role role);
 }
