+++ daengle-persistence/src/main/java/ddog/persistence/port/CareEstimatePersist.java
@@ -8,4 +8,6 @@ public interface CareEstimatePersist {
     void save(CareEstimate careEstimate);
     CareEstimate getByCareEstimateId(Long careEstimateId);
     List<CareEstimate> findCareEstimatesByPetId(Long petId);
+    List<CareEstimate> findGeneralCareEstimates(String address);
+    List<CareEstimate> findDesignationCareEstimates(Long vetId);
 }
