+++ daengle-persistence/src/main/java/ddog/persistence/port/GroomerPersist.java
@@ -3,6 +3,5 @@ package ddog.persistence.port;
 import ddog.domain.groomer.Groomer;
 
 public interface GroomerPersist {
-
     Groomer getGroomerByAccountId(Long accountId);
 }
