+++ daengle-persistence/src/main/java/ddog/persistence/port/PetPersist.java
@@ -3,8 +3,6 @@ package ddog.persistence.port;
 import ddog.domain.pet.Pet;
 
 public interface PetPersist {
-
     Pet save(Pet pet);
-
     void deletePetById(Long petId);
 }
