+++ daengle-persistence/src/main/java/ddog/persistence/port/UserPersist.java
@@ -3,10 +3,7 @@ package ddog.persistence.port;
 import ddog.domain.user.User;
 
 public interface UserPersist {
-
     void save(User user);
-
     User findByAccountId(Long accountId);
-
     Boolean hasNickname(String nickname);
 }
