+++ daengle-persistence/src/main/java/ddog/persistence/port/VetPersist.java
@@ -1,9 +1,7 @@
-package ddog.vet.application;
+package ddog.persistence.port;
 
 import ddog.domain.vet.Vet;
 
-public interface VetRepository {
-
+public interface VetPersist {
     Vet getVetByAccountId(Long accountId);
-
 }
