+++ daengle-vet-api/src/main/java/ddog/vet/application/exception/common/CustomError.java
@@ -1,4 +1,4 @@
-package ddog.vet.exception.common;
+package ddog.vet.application.exception.common;
 
 import lombok.Data;
 import org.springframework.http.HttpStatus;
