+++ daengle-vet-api/src/main/java/ddog/vet/application/exception/common/IllegalParameterException.java
@@ -1,4 +1,4 @@
-package ddog.vet.exception.common;
+package ddog.vet.application.exception.common;
 
 public class IllegalParameterException extends RuntimeException {
     public IllegalParameterException(String message, Object... args) {
