+++ daengle-vet-api/src/main/java/ddog/vet/application/exception/common/NotFoundException.java
@@ -1,4 +1,4 @@
-package ddog.vet.exception.common;
+package ddog.vet.application.exception.common;
 
 import org.springframework.http.HttpStatus;
